==> notebooks/4.5_final_processing.ipynb <==
# Bank Customer Churn - Final Processing

This notebook serves just as a sketch of the data processing step, so the inference notebook may not be poluted. Every code here was already discussed in the previous notebooks.

import pandas as pd

def processing_pipeline(df: pd.DataFrame) -> pd.DataFrame:
    df_copy = df.copy()
    df_copy = df_copy.iloc[:, 3:]
    
    df_copy['Gender'].replace({'Male': 0, 'Female': 1}, inplace=True)
    geography_dummies = pd.get_dummies(df_copy['Geography']).astype(int)
    df_copy = pd.concat([df_copy, geography_dummies],axis=1).drop('Geography', axis=1)
    df_copy['Products_Tenure_relation'] = df_copy['NumOfProducts'] * df_copy['Tenure']
    
    return df_copy

df_test = pd.read_csv('../data/raw/Abandono_teste.csv', sep=';')

df_test_processed = processing_pipeline(df_test)

df_test_processed.to_csv('../data/processed/inference_data.csv', index=False)

df_test
