==> notebooks/5.0_lc_inference.ipynb <==
# Bank Customer Churn - Inference

Here it is presented the final step of the pipeline: inference.

As simple as it seems, it is necessary to design it to be robust against problems like different models in the same folder, new data with different schema of the one the model was trained, and assert that the output is in the expected format to be served to an API, for example.

In this case, due to the simplicity of the project, there's no mistery here in what to be done, but it is useful to think about it as part of a fully served pipeline.



import os
import pickle
import re

import pandas as pd
import numpy as np

from xgboost import XGBClassifier

def obtain_score_from_name(file_name: str) -> float | None:
    regex_pattern = r'_(\d+\.\d+)\.pkl$'
    match = re.match(regex_pattern, file_name)
    if match:
        score = float(match.group('score'))
        return score
    
def obtain_model_with_higher_score(file_paths: list[str]) -> str:
    def sorting_function(x: str)->float:
        score = obtain_score_from_name(x)
        if score is None:
            return float('inf')
        return score
        
    sorted_paths = sorted(file_paths, key=sorting_function)
    return sorted_paths[0]

model_folder_path = '../models'
files = os.listdir(model_folder_path)
best_model_file = obtain_model_with_higher_score(files)

with open(os.path.join(model_folder_path,best_model_file), 'rb') as f:
    xgb_model = pickle.load(f)

df_test = pd.read_csv('../data/raw/Abandono_teste.csv', sep=';')
df_inference = pd.read_csv('../data/processed/inference_data.csv')


df_inference

model_preds = xgb_model.predict(df_inference)
df_test['Exited'] = model_preds

predictions = df_test[['RowNumber', 'Exited']]
predictions

With this, we finish the reasoning and design of the bank customer churn project. What's left now is the construction of the pipeline code - that must be finished when you read this notebook - adapting the code presented to good software engineering practices in order to mantain it clean, easy to maintain and reproducible.
